Add HTTP interface and server build target
Kaniko is normally packaged as a binary that receives build arguments,
makes a container image and pushes it to a registry.

Running the kaniko binary outside of its container is not supported so
it's difficult to invoke it from other containers.

This change adds a simple server interface to control Kaniko builds
over HTTP. The Kaniko server image can be run as a sidecar next to
other containers which can share context through volumens and can
invoke it through HTTP calls.